Move stdio functions to the separate files.
