+++ Library/stdio/clearerr.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+void clearerr(FILE *stream)
+{
+    stream->error = 0;
+}
