+++ Library/stdio/fclose.c
@@ -1,0 +1,9 @@
+#include "../stdio.h"
+
+int fclose(FILE *stream)
+{
+    fflush(stream);
+    free(stream);
+
+    return 0;
+}
