+++ Library/stdio/feof.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int feof(FILE *stream)
+{
+    return stream->pos < stream->limit;
+}
