+++ Library/stdio/ferror.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int ferror(FILE *stream)
+{
+    return stream->error;
+}
