+++ Library/stdio/fgetpos.c
@@ -1,0 +1,7 @@
+#include "../stdio.h"
+
+int fgetpos(FILE *stream, fpos_t *pos)
+{
+    *pos = stream->pos;
+    return 0;
+}
