+++ Library/stdio/fputc.c
@@ -1,0 +1,7 @@
+#include "../stdio.h"
+
+int fputc(int character, FILE *stream)
+{
+    fwrite(&character, 1, 1, stream);
+    return character;
+}
