+++ Library/stdio/fputs.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int fputs(const char *str, FILE *stream)
+{
+    return fwrite(str, strlen(str), 1, stream);
+}
