+++ Library/stdio/fsetpos.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int fsetpos(FILE *stream, const fpos_t *pos)
+{
+    return fseek(stream, *pos, SEEK_SET);
+}
