+++ Library/stdio/ftell.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+long int ftell(FILE *stream)
+{
+    return stream->pos;
+}
