+++ Library/stdio/getc.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int getc(FILE *stream)
+{
+    return fgetc(stream);
+}
