+++ Library/stdio/getchar.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int getchar()
+{
+    return fgetc(stdin);
+}
