+++ Library/stdio/gets.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+char *gets(char *str)
+{
+    return fgets(str, INT32_MAX, stdin);
+}
