+++ Library/stdio/putc.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int putc(int character, FILE *stream)
+{
+    return fputc(character, stream);
+}
