+++ Library/stdio/putchar.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int putchar(int character)
+{
+    return fputc(character, stdout);
+}
