+++ Library/stdio/puts.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+int puts(const char *str)
+{
+    return fputs(str, stdout);
+}
