+++ Library/stdio/rewind.c
@@ -1,0 +1,6 @@
+#include "../stdio.h"
+
+void rewind(FILE *stream)
+{
+    fseek(stream, 0, SEEK_SET);
+}
